requirements.txt
# requirements.txt
# ============================================
# CORE - APPLICATION WEB
# ============================================
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# ============================================
# IA & APIS
# ============================================
openai==1.12.0
requests==2.31.0
facebook-sdk==3.1.0
unsplash==2.1.0

# ============================================
# GOOGLE SHEETS & DATA
# ============================================
gspread==5.12.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2

# ============================================
# SYSTÈME AUTOMATIQUE
# ============================================
schedule==1.2.0
threading==0.0.0

# ============================================
# HEROKU / PYTHONANYWHERE
# ============================================
psycopg2-binary==2.9.9  # Pour PostgreSQL si besoin